# src/generated/swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"LoginResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"token": {
					"type": "string"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
		"LoginRequest": {
			"properties": {
				"password": {
					"type": "string"
				}
			},
			"required": [
				"password"
			],
			"type": "object",
			"additionalProperties": false
		},
		"ContactResponse": {
			"properties": {
				"id": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"subject": {
					"type": "string"
				},
				"message": {
					"type": "string"
				},
				"status": {
					"type": "string"
				},
				"source": {
					"type": "string"
				},
				"discoveryDetails": {
					"additionalProperties": true
				},
				"createdAt": {
					"type": "string"
				},
				"updatedAt": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"name",
				"email",
				"subject",
				"message",
				"status",
				"source",
				"createdAt",
				"updatedAt"
			],
			"type": "object",
			"additionalProperties": false
		},
		"CreateContactResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"data": {
					"$ref": "#/definitions/ContactResponse"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
		"CreateContactRequest": {
			"properties": {
				"name": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"subject": {
					"type": "string"
				},
				"message": {
					"type": "string"
				}
			},
			"required": [
				"name",
				"email",
				"subject",
				"message"
			],
			"type": "object",
			"additionalProperties": false
		},
		"ContactsListResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"data": {
					"items": {
						"$ref": "#/definitions/ContactResponse"
					},
					"type": "array"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
		"SimpleResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"error": {
					"type": "string"
				}
			},
			"required": [
				"success"
			],
			"type": "object",
			"additionalProperties": false
		},
		"UpdateContactStatusRequest": {
			"properties": {
				"status": {
					"type": "string",
					"enum": [
						"new",
						"read",
						"replied"
					],
					"x-nullable": false
				}
			},
			"required": [
				"status"
			],
			"type": "object",
			"additionalProperties": false
		},
...
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"description": "Prisma, TypeScript, Swagger & tsoa Backend",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/contacts": {
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Contact created successfully",
						"schema": {
							"$ref": "#/definitions/CreateContactResponse"
						}
					},
					"400": {
						"description": "Missing required fields"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Public — submit the website's \"Contact Us\" form.",
				"tags": [
					"Contacts"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CreateContactRequest"
						}
					}
				]
			},
			"get": {
				"operationId": "List",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Contacts retrieved successfully",
						"schema": {
							"$ref": "#/definitions/ContactsListResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"description": "Get all contacts (admin only)",
				"tags": [
					"Contacts"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/contacts/login": {
			"post": {
				"operationId": "Login",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Logged in successfully",
						"schema": {
							"$ref": "#/definitions/LoginResponse"
						}
					},
					"401": {
						"description": "Incorrect password"
					},
					"500": {
						"description": "Server auth not configured"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Admin login - get JWT token",
				"tags": [
					"Contacts"
				],
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/LoginRequest"
						}
					}
				]
			}
		},
		"/api/contacts/{id}/status": {
			"put": {
				"operationId": "UpdateStatus",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Status updated successfully",
						"schema": {
							"$ref": "#/definitions/SimpleResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Contact not found"
					}
				},
				"consumes": [
					"application/json"
				],
				"description": "Update contact status (admin only)",
				"tags": [
					"Contacts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "request",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UpdateContactStatusRequest"
						}
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"type": "string"
					}
				]
			}
		},
		"/api/contacts/{id}": {
			"delete": {
				"operationId": "Remove",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Contact deleted successfully",
						"schema": {
							"$ref": "#/definitions/SimpleResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"404": {
						"description": "Contact not found"
					}
				},
				"description": "Delete a contact (admin only)",
				"tags": [
					"Contacts"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string"
					},
					{
						"in": "header",
						"name": "Authorization",
						"required": false,
						"type": "string"
					}
				]
			}
		},
...
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"jwt": {
			"type": "apiKey",
			"name": "Authorization",
			"in": "header"
		}
	}
}